[i-43] Create lib/types for shared types
